pyproject.toml
[tool.poetry]
name = "notion-semantic-search"
version = "0.1.0"
description = ""
authors = ["Jeremy Udit <jcudit@github.com>"]
packages = [
    { include="embed", from="." },
    { include="server", from="." },
]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "7.1.1"
click = "8.0.4"
jsonlines = "3.1.0"
tqdm = "4.64.0"
fastapi = "0.68.0,<0.69.0"
pydantic = "^1.9.2"
uvicorn = "^0.18.3"
requests-ratelimiter = "^0.3.2"
notion-client = "^1.0.0"
cohere = "^2.2.3"
numpy = "^1.23.2"
faiss-cpu = "^1.7.2"

[tool.poetry.dev-dependencies]
yapf = "^0.32.0"
toml = "^0.10.2"
flake8 = "^5.0.4"
isort = "^5.10.1"
autoflake = "^1.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
embed = "embed.cli:cli"
search = "search.cli:cli"

